+++ packages/headless/src/api/search/search-api-client.test.ts
@@ -67,6 +67,7 @@ describe('search api client', () => {
         q: state.query.q,
         cq: '',
         aq: '',
+        debug: false,
         numberOfResults: state.pagination.numberOfResults,
         sortCriteria: state.sortCriteria,
         firstResult: state.pagination.firstResult,