+++ packages/headless/src/api/search/search/ranking-expression.ts
@@ -1,0 +1,4 @@
+export interface RankingExpression {
+  expression: string;
+  modifier: string;
+}