+++ packages/headless/src/controllers/facets/_common/facet-option-definitions.ts
@@ -12,7 +12,9 @@ import {FacetSearchRequestOptions} from '../../../features/facets/facet-search-s
  * A unique identifier for the controller.
  * By default, a unique random identifier is generated.
  */
-export const facetId = new StringValue();
+export const facetId = new StringValue({
+  regex: /^[a-zA-Z0-9-_]+$/,
+});
 /** The field whose values you want to display in the facet.*/
 export const field = new StringValue({required: true});
 export const basePath = new ArrayValue({